mv test from unit to integration for connections
