+++ src/test/unit/indexer/indexer.test.ts
@@ -1,33 +1,8 @@
 import { expect } from 'chai'
 import { stub } from 'sinon'
 import { describe, it } from 'mocha'
-import {
-  INDEXER_CRAWLING_EVENT_EMITTER,
-  OceanIndexer
-} from '../../../components/Indexer/index.js'
+import { OceanIndexer } from '../../../components/Indexer/index.js'
 import { RPCS } from '../../../@types/blockchain'
-import {
-  OverrideEnvConfig,
-  buildEnvOverrideConfig,
-  getMockSupportedNetworks,
-  setupEnvironment
-} from '../../utils/utils.js'
-import { OceanNodeConfig } from '../../../@types/OceanNode.js'
-import { Database } from '../../../components/database/index.js'
-import { OceanNode } from '../../../OceanNode.js'
-import {
-  ENVIRONMENT_VARIABLES,
-  INDEXER_CRAWLING_EVENTS
-} from '../../../utils/constants.js'
-import { getConfiguration } from '../../../utils/config.js'
-import { Blockchain } from '../../../utils/blockchain.js'
-import {
-  getNetworkHeight,
-  getDeployedContractBlock
-} from '../../../components/Indexer/utils.js'
-import { homedir } from 'os'
-import { DEVELOPMENT_CHAIN_ID } from '../../../utils/address.js'
-import { sleep } from '../../../utils/util.js'
 
 class MockDatabase {
   indexer = {
@@ -80,59 +55,5 @@ describe('OceanIndexer', () => {
       .false
     // eslint-disable-next-line no-unused-expressions
     expect(mockWorker.on.calledThrice).to.be.false
-  })
-})
-
-describe('OceanIndexer - crawler threads', () => {
-  let envOverrides: OverrideEnvConfig[]
-  let config: OceanNodeConfig
-  let db: Database
-  let oceanNode: OceanNode
-  let oceanIndexer: OceanIndexer
-  let blockchain: Blockchain
-
-  const mockSupportedNetworks: RPCS = getMockSupportedNetworks()
-  const chainID = DEVELOPMENT_CHAIN_ID.toString()
-  mockSupportedNetworks[chainID].startBlock = 2
-
-  before(async () => {
-    envOverrides = buildEnvOverrideConfig(
-      [ENVIRONMENT_VARIABLES.RPCS, ENVIRONMENT_VARIABLES.ADDRESS_FILE],
-      [
-        JSON.stringify(mockSupportedNetworks),
-        `${homedir}/.ocean/ocean-contracts/artifacts/address.json`
-      ]
-    )
-    envOverrides = await setupEnvironment(null, envOverrides)
-    config = await getConfiguration(true)
-    db = await new Database(config.dbConfig)
-
-    oceanNode = OceanNode.getInstance(db)
-    oceanIndexer = new OceanIndexer(db, mockSupportedNetworks)
-    blockchain = new Blockchain(
-      mockSupportedNetworks[chainID].rpc,
-      mockSupportedNetworks[chainID].chainId
-    )
-
-    oceanNode.addIndexer(oceanIndexer)
-  })
-  it('should start a worker thread and handle RPCS "startBlock"', async () => {
-    const netHeight = await getNetworkHeight(blockchain.getProvider())
-    const deployBlock = getDeployedContractBlock(mockSupportedNetworks[chainID].chainId)
-    const { indexer } = db
-    const updatedIndex = await indexer.update(Number(chainID), 1)
-
-    expect(updatedIndex.lastIndexedBlock).to.be.equal(1)
-
-    await oceanIndexer.startThreads()
-    INDEXER_CRAWLING_EVENT_EMITTER.addListener(
-      INDEXER_CRAWLING_EVENTS.CRAWLING_STARTED,
-      (data: any) => {
-        const { startBlock, deployedContractBlock, networkHeight } = data
-        expect(startBlock).to.be.equal(1)
-        expect(deployedContractBlock).to.be.equal(deployBlock)
-        expect(networkHeight).to.be.equal(netHeight)
-      }
-    )
   })
 })
